moved tests to correct place, and refactored to use factories and conftest.py
